Remove dbus import from non-linux files
