+++ plugins/inputs/systemd_units/systemd_units_notlinux.go
@@ -4,6 +4,8 @@ package systemd_units
 
 import "github.com/influxdata/telegraf"
 
+type client interface{}
+
 func (s *SystemdUnits) Init() error {
 	s.Log.Info("Skipping plugin as it is not supported by this platform!")
